Implement Mapper000 class with CPU and PPU read/write methods

+++ Mapper.h
@@ -10,10 +10,10 @@ class Mapper{
         Mapper(uint8_t prgBanks, uint8_t chrBanks);
         ~Mapper();
     private:
-        virtual bool modCpuRead(uint16_t addr, uint16_t &mappedAddr);
-        virtual bool modCpuWrite(uint16_t addr, uint16_t &mappedAddr);
-        virtual bool modPpuRead(uint16_t addr, uint16_t &mappedAddr);
-        virtual bool modPpuWrite(uint16_t addr, uint16_t &mappedAddr);
+        virtual bool modCpuRead(uint16_t addr, uint32_t &mappedAddr);
+        virtual bool modCpuWrite(uint16_t addr, uint32_t &mappedAddr);
+        virtual bool modPpuRead(uint16_t addr, uint32_t &mappedAddr);
+        virtual bool modPpuWrite(uint16_t addr, uint32_t &mappedAddr);
     protected:
         uint8_t numPrgBanks;
         uint8_t numChrBanks;
